vm: fix MSVC warnings

+++ vm/image.cpp
@@ -229,7 +229,7 @@ FILE* factor_vm::open_image(vm_parameters *p)
 			std::cout << strerror(errno) << std::endl;
 			exit(1);
 		}
-		safe_fseek(file, -sizeof(embedded_image_footer), SEEK_END);
+		safe_fseek(file, -(off_t)sizeof(embedded_image_footer), SEEK_END);
 		embedded_image_footer footer;
 		safe_fread(&footer, sizeof(embedded_image_footer), 1, file);
 		if (footer.magic != image_magic)
@@ -237,7 +237,7 @@ FILE* factor_vm::open_image(vm_parameters *p)
 			std::cout << "No embedded image" << std::endl;
 			exit(1);
 		}
-		safe_fseek(file, footer.image_offset, SEEK_SET);
+		safe_fseek(file, (off_t)footer.image_offset, SEEK_SET);
 		return file;
 	}
 	else
@@ -372,9 +372,9 @@ bool factor_vm::embedded_image_p()
 	FILE *file = OPEN_READ(vm_path);
 	if (!file)
 		return false;
-	safe_fseek(file, -sizeof(embedded_image_footer), SEEK_END);
+	safe_fseek(file, -(off_t)sizeof(embedded_image_footer), SEEK_END);
 	embedded_image_footer footer;
-	safe_fread(&footer, sizeof(embedded_image_footer), 1, file);
+	safe_fread(&footer, (off_t)sizeof(embedded_image_footer), 1, file);
 	fclose(file);
 	return footer.magic == image_magic;
 }
